+++ integrated_channels/integrated_channel/models.py
@@ -2,24 +2,17 @@
 Database models for Enterprise Integrated Channel.
 """
 from __future__ import absolute_import, unicode_literals
+import logging
 
 from django.db import models
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import ugettext_lazy as _
 from model_utils.models import TimeStampedModel
 
-from enterprise.models import EnterpriseCustomer, EnterpriseCourseEnrollment
-from enterprise.utils import NotConnectedToOpenEdX
+from enterprise.models import EnterpriseCustomer
 
-try:
-    from certificates.api import GeneratedCertificate
-except ImportError:
-    GeneratedCertificate = None
 
-try:
-    from opaque_keys.edx.keys import CourseKey
-except ImportError:
-    CourseKey = None
+LOGGER = logging.getLogger(__name__)
 
 
 @python_2_unicode_compatible
@@ -79,67 +72,34 @@ class EnterpriseCustomerPluginConfiguration(TimeStampedModel):
         """
         Returns an capitalized identifier for this channel class, unique among subclasses.
         """
-        raise NotImplementedError(_('Implemented in concrete subclass.'))
+        raise NotImplementedError('Implemented in concrete subclass.')
 
-    def transmit_learner_data(self):
+    def get_learner_data_record(self, enterprise_enrollment, completed_date=None, grade=None):
         """
-        Collect and transmit learner data for the ``EnterpriseCustomer``.
+        Returns an capitalized identifier for this channel class, unique among subclasses.
         """
-        transmitter = self.get_learner_data_transmitter()
-        for learner_data in self.collect_learner_data():
-            transmitter.transmit(learner_data)
+        raise NotImplementedError('Implemented in concrete subclass.')
 
-    def collect_learner_data(self):
+    def get_learner_data_exporter(self, user):
         """
-        Collect learner data for the ``EnterpriseCustomer`` where data sharing consent is granted.
-
-        Yields a learner data object for each enrollment, containing:
-
-        * ``enterprise_enrollment``: ``EnterpriseCourseEnrollment`` object.
-        * ``certificate``: ``GeneratedCertificate`` for the user+course; None if not found.
-          "Course completion" occurs when course certificates are issued.
+        Returns the class that can serialize the learner course completion data to the integrated channel.
         """
-
-        if CourseKey is None or GeneratedCertificate is None:
-            raise NotConnectedToOpenEdX(_('This package must be installed in an OpenEdX environment.'))
-
-        # Fetch the consenting enrollment data, including the enterprise_customer_user
-        enrollment_queryset = EnterpriseCourseEnrollment.objects.select_related(
-            'enterprise_customer_user'
-        ).filter(
-            enterprise_customer_user__enterprise_customer=self.enterprise_customer,
-        )
-        for enterprise_enrollment in enrollment_queryset:
-
-            # Omit any enrollments where consent has not been granted
-            if not enterprise_enrollment.consent_available():
-                continue
-
-            # Fetch the user+course certificate, if found
-            try:
-                certificate = GeneratedCertificate.eligible_certificates.get(
-                    user__id=enterprise_enrollment.enterprise_customer_user.user_id,
-                    course_id=CourseKey.from_string(enterprise_enrollment.course_id),
-                )
-            except GeneratedCertificate.DoesNotExist:
-                certificate = None
-
-            yield self.get_learner_data(
-                enterprise_enrollment=enterprise_enrollment,
-                certificate=certificate,
-            )
-
-    def get_learner_data(self, enterprise_enrollment, certificate):
-        """
-        Returns the class that can serialize the learner completion data to the integrated channel.
-        """
-        raise NotImplementedError(_('Implemented in concrete subclass.'))
+        raise NotImplementedError('Implemented in concrete subclass.')
 
     def get_learner_data_transmitter(self):
         """
-        Returns the class that can transmit the learner completion data to the integrated channel.
+        Returns the class that can transmit the learner course completion data to the integrated channel.
         """
         raise NotImplementedError("Implemented in concrete subclass.")
+
+    def transmit_learner_data(self, user):
+        """
+        Iterate over each learner data record and transmit it to the integrated channel.
+        """
+        exporter = self.get_learner_data_exporter(user)
+        transmitter = self.get_learner_data_transmitter()
+        for learner_data in exporter.collect_learner_data():
+            transmitter.transmit(learner_data)
 
     def get_course_data_exporter(self, user):
         """